requirements.txt
gpiozero
lgpio
RPi.GPIO
wand
pillow
spidev
openai
